pyproject.toml
[tool.poetry]
name = "sender"
version = "0.1.0"
description = ""
authors = ["Maksim <maksim.malhinau@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.5"
djangorestframework = "^3.16.0"
djangorestframework-simplejwt = "^5.5.1"
django-redis = "^6.0.0"
django-celery-results = "^2.6.0"
celery = "^5.5.3"
django-celery-beat = "^2.8.1"
python-dotenv = "^1.1.1"
drf-spectacular = "^0.28.0"
django-split-settings = "^1.3.2"
psycopg2 = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
